requirements.txt
annotated-types==0.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
progress==1.6
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
